% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLSestimates.R
\name{beta.estimator}
\alias{beta.estimator}
\title{OLS Beta Estimates using either Cholesky or QR Decomposition}
\usage{
beta.estimator(X, y, W)
}
\arguments{
\item{X:}{An n by p design matrix (including intercept)}

\item{y:}{A numeric vector of length n}

\item{W:}{A logical (boolean) object. If W = T, then OLS algorithm uses Cholesky, otherwise OLS algorithm uses QR}
}
\value{
A numeric vector
}
\description{
OLS Beta Estimates using either Cholesky or QR Decomposition
}
\examples{

set.seed(1247)
X <- matrix(sample(1:20, 16, replace=T), ncol=4)
y <- sample(1:20, 4, replace=T)
W = T
beta.estimator(X, y, W)

}
