% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.beta.est.R
\name{chol.beta.estimator}
\alias{chol.beta.estimator}
\title{OLS Beta Estimates using Cholesky Decomposition}
\usage{
\method{chol}{beta.estimator}(X, y)
}
\arguments{
\item{X:}{An n by p design matrix (including intercept)}

\item{y:}{A numeric vector of length n}
}
\value{
A numeric vector
}
\description{
OLS Beta Estimates using Cholesky Decomposition
}
\examples{

X <- matrix(c(4, 0, 0, 0, 3, 7, 0, 0, 7, 8, 5, 0, 9, 2, 11, 15), byrow=T, nrow=4)
y <- c(7, 18, 4, 11)
chol.beta.estimator(X, y)

}
