% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLSestimates.R
\name{chol.beta.estimator}
\alias{chol.beta.estimator}
\title{OLS Beta Estimates using either Cholesky or QR Decomposition}
\usage{
\method{chol}{beta.estimator}(X, y, W)
}
\arguments{
\item{X:}{An n by p design matrix (including intercept)}

\item{y:}{A numeric vector of length n}

\item{W:}{A logical (boolean) object. If W = T, then OLS algorithm uses Cholesky, otherwise OLS algorithm uses QR}
}
\value{
A numeric vector
}
\description{
OLS Beta Estimates using either Cholesky or QR Decomposition
}
\examples{

sim.data <- read.csv("/Users/dmwakima/Documents/09_R/07_Winter_23/STATS230/firstpkg/homework2_regression.csv")
source("/Users/dmwakima/Documents/09_R/07_Winter_23/STATS230/firstpkg/R/matvecprod.R")
dim(sim.data)
head(sim.data)
X <- as.matrix(sim.data[, c(3:6)])
y <- as.vector(sim.data[, 1])
chol.beta.estimator(X, y)

}
