% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMfunction.R
\name{EMfunction}
\alias{EMfunction}
\title{MLE of Normal Mixture Models using the E-M Algorithm}
\usage{
EMfunction(y, theta, iter)
}
\arguments{
\item{y:}{a vector of observed data}

\item{theta:}{a vector of initial values for all the parameters}

\item{iter:}{an integer for the number of iterations of the optimization task}
}
\value{
a vector of the Maximum Likelihood Estimates of your parameters
}
\description{
MLE of Normal Mixture Models using the E-M Algorithm
}
\examples{
rm(list=ls())
test_data <- read.table("mixture_data.txt", header = F)
y <- as.matrix(test_data)
theta = c(0.5, 4, 8)
iter = 200
EMfunction(y, theta, iter)

}
