% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaMCMC.R
\name{lnorm_rw_next}
\alias{lnorm_rw_next}
\title{Rosenbluth-Hastings Gamma MCMC}
\usage{
lnorm_rw_next(curr_val, tuning_param)
}
\arguments{
\item{curr_val}{This is the current state of the R-H Markov chain}

\item{tuning_param}{This is a tuning parameter of the Normal random walk}
}
\value{
Numeric vector of length 2. The first element of the vector contains the next state of the R-H Markov chain. 
The second element contains 0 if the proposed values was rejected and 1 otherwise.
}
\description{
Rosenbluth-Hastings Gamma MCMC
}
\examples{
lnorm_rw_next(1, 1.5)
}
