% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_MLE.R
\name{logistic_MLE}
\alias{logistic_MLE}
\title{Logistic MLE}
\usage{
logistic_MLE(Y, X, method)
}
\arguments{
\item{method}{}

\item{Y:}{n times 1 matrix of binary responses}

\item{X:}{n times p design matrix}
}
\value{
a vector of MLE estimates of a logistic regression model
}
\description{
Logistic MLE
}
