% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matvecprod.R
\name{matvecprod}
\alias{matvecprod}
\title{Products of Vectors and Matrices}
\usage{
matvecprod(A, B, V, W)
}
\arguments{
\item{A}{a m by n matrix}

\item{B}{a n by p matrix}

\item{V}{a vector}

\item{W}{logical (T or F), if T we compute (AB)v if F we compute A(Bv).}
}
\value{
A numeric vector of length m
}
\description{
This package computes the product of two matrices and a vector. 
This can be done by the user in two ways by switching the fourth argument.
One way computes (AB)v and the other way computes A(Bv). 
In the example below we use micro-benching to measure the performance of the two ways to compute the product above.
}
\examples{
A = matrix(data=c(1,2,3,4,2,1,9,2), nrow = 4, ncol=2, byrow=T)
B = matrix(data=c(3,4,5,6), nrow =2, ncol=2, byrow=T)
V = c(3,8)
W = 1
matvecprod(A, B, V, W)
bench::mark(matvecprod(A, B, V,W=1), matvecprod(A, B, V,W=2))
}
