% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.beta.est.R
\name{QR.beta.estimator}
\alias{QR.beta.estimator}
\title{OLS Estimates using QR Decomposition}
\usage{
QR.beta.estimator(X, y)
}
\arguments{
\item{X:}{A n by p design matrix (including intercept)}

\item{y:}{A numeric vector of length n}
}
\value{
A numeric vector
}
\description{
OLS Estimates using QR Decomposition
}
\examples{

X <- matrix(c(16, 12, 28, 36, 12, 58, 77, 41, 28, 77, 138, 134, 36, 41, 134, 431), byrow =T, nrow=4)
y <- c(7, 18, 4, 11)
QR.beta.estimator(X, y)

}
