% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLSestimates.R
\name{beta.estimator}
\alias{beta.estimator}
\title{OLS Beta Estimates using either Cholesky or QR Decomposition}
\usage{
beta.estimator(X, y, W)
}
\arguments{
\item{X:}{An n by p design matrix (including intercept)}

\item{y:}{A numeric vector of length n}

\item{W:}{A logical (boolean) object. If W = T, then OLS algorithm uses Cholesky, otherwise OLS algorithm uses QR}
}
\value{
A numeric vector
}
\description{
OLS Beta Estimates using either Cholesky or QR Decomposition
}
\examples{

sim.data <- read.csv("https://github.com/miningroup/your-stat-230-r-package-davymwax/blob/main/homework2_regression.csv")
source("https://github.com/miningroup/your-stat-230-r-package-davymwax/blob/main/R/matvecprod.R")
dim(sim.data)
head(sim.data)
X <- as.matrix(sim.data[, c(3:6)])
y <- as.vector(sim.data[, 1])
W = T
beta.estimator(X, y, W)

}
