% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matvecprod.R
\name{matvecprod}
\alias{matvecprod}
\title{Matrix and Vector Products}
\usage{
matvecprod(A, B, V, W)
}
\arguments{
\item{A}{A matrix}

\item{B}{A matrix}

\item{V}{A vector}

\item{W}{A number 1 or 2, which tells us the way the product is to be computed}
}
\value{
A numeric vector
}
\description{
Matrix and Vector Products
}
\examples{
A = matrix(data=c(1,2,3,4), nrow =2, ncol=2, byrow=T)
B = matrix(data=c(3,4,5,6), nrow =2, ncol=2, byrow=T)
V = c(3,8)
V = c(3,8)
W = 1
matvecprod(A, B, V, W)
}
