% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multvnorm.R
\name{multvnorm}
\alias{multvnorm}
\title{Simulating a Multivariate Normal}
\usage{
multvnorm(mu, sigma, N)
}
\arguments{
\item{mu}{: a numeric vector}

\item{sigma}{: an n by n positive definite matrix}

\item{N}{: numeric integer greater than or equal 1.}
}
\value{
a list
}
\description{
This package simulates N realizations multivariate normal given
}
\examples{
N = 100
n = 4
p = 2
mu = rep(0, n)
X = matrix(data=sample(1:20, n*p), nrow=n)
sigma = cov(X)
multvnorm(mu, sigma, N)
}
