% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multvnorm.R
\name{multvnorm}
\alias{multvnorm}
\title{Simulating a Multivariate Normal}
\usage{
multvnorm(mu, sigma, N)
}
\arguments{
\item{mu}{}

\item{sigma}{}

\item{N}{}
}
\description{
This package simulates N realizations multivariate normal given
}
\examples{
N = 100
mu = rep(0, n)
X = matrix(data=sample(1:20, n*n), ncol=n)
sigma = X \%*\% t(X)
}
